feat(Candidate Module) -> Add Candidate module Containing Candidate Profile into Dev

+++ backend/Api/Module/Jobs/Jobs.Application/JobApplication/Create/v1/CreateJobApplicationHandler.cs
@@ -14,8 +14,8 @@ public sealed class CreateJobApplicationHandler(
     {
         ArgumentNullException.ThrowIfNull(request);
         var jobApplication = Domain.JobApplication.Create(
-            (int)request.JobId!,
-            (int)request.CandidateId!,
+            request.JobId!,
+            request.CandidateId!,
             request.CoverLetter!
         );
         await repository.AddAsync(jobApplication, cancellationToken);
